requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.32.3
trafilatura>=2.0.0
python-dotenv>=1.0.1
pytest>=8.3.4
aiohttp>=3.11.12
flask_cors>=5.0.0
pytest-asyncio>=0.25.3
pytest-mock>=3.14.0